backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0

python-dotenv==1.2.1
python-multipart==0.0.21

groq==1.0.0

langchain==1.2.0
langchain-core==1.2.6
langserve==0.3.3

sentence-transformers==5.2.0
torch==2.9.1
transformers==4.57.3

scikit-learn==1.8.0
scipy==1.16.3
numpy==1.26.4

pypdf==6.5.0
tqdm==4.67.1
requests==2.32.5
PyYAML==6.0.3
tenacity==9.1.2
sse-starlette==1.8.2
